refactor(core): Extract AST validation to dedicated class
Extract AST validation logic from TemplateSandbox to new ASTValidator
class to comply with ESLint max-lines rule (300 lines limit).

Changes:
- Create ASTValidator.ts with all AST validation logic (243 lines)
- Update TemplateSandbox.ts to use ASTValidator instance (273 lines, was 330)
- Maintain same validation behavior with cleaner separation of concerns

Fixes ESLint violation without disabling linting rules.
All tests pass (214 tests in template suite).
